refactor: useAvailableWidth
