+++ packages/react-responsive-pagination/src/hooks/useAvailableWidth.ts
@@ -1,5 +1,17 @@
+import { useState } from 'react';
 import { useContentWidth } from './useContentWidth.js';
 
-export function useAvailableWidth(element: Element | null) {
-  return useContentWidth(element?.parentElement ?? undefined);
+export function useAvailableWidth(overrideWidth: number | undefined) {
+  const [element, setElement] = useState<Element | null>(null);
+
+  const parentElement = element?.parentElement ?? undefined;
+
+  const width = useContentWidth(
+    overrideWidth === undefined ? parentElement : undefined,
+  );
+
+  return {
+    width: overrideWidth ?? width,
+    ref: setElement,
+  };
 }
